+++ rasa_core/channels/telegram.py
@@ -103,9 +103,12 @@ class TelegramInput(HttpInputComponent):
             s = out_channel.setWebhook(self.webhook_url)
             if s:
                 return "Webhook setup successful"
+                logger.info("Webhook Setup Successful")
             else:
                 return "Invalid webhook"
-
+                logger.warn("Webhook Setup Failed")
+        set_webhook()
+        
         @telegram_webhook.route("/webhook", methods=['GET', 'POST'])
         def message():
             if request.method == 'POST':