+++ rasa_core/policies/trainer.py
@@ -20,14 +20,14 @@ class PolicyTrainer(object):
         self.ensemble = ensemble
         self.featurizer = featurizer
 
-    def train(self, filename=None, max_history=3,
+    def train(self, resource_name=None, max_history=3,
               augmentation_factor=20, max_training_samples=None,
               max_number_of_trackers=2000, remove_duplicates=True, **kwargs):
         """Trains a policy on a domain using training data from a file.
 
         :param augmentation_factor: how many stories should be created by
                                     randomly concatenating stories
-        :param filename: story file containing the training conversations
+        :param resource_name: story file containing the training conversations
         :param max_history: number of past actions to consider for the
                             prediction of the next action
         :param max_training_samples: specifies how many training samples to
@@ -44,25 +44,25 @@ class PolicyTrainer(object):
         logger.debug("Policy trainer got kwargs: {}".format(kwargs))
         check_domain_sanity(self.domain)
 
-        training_data = self._prepare_training_data(filename, max_history,
-                                           augmentation_factor,
-                                           max_training_samples,
-                                           max_number_of_trackers,
-                                           remove_duplicates)
+        training_data = self._prepare_training_data(
+                resource_name, max_history, augmentation_factor,
+                max_training_samples, max_number_of_trackers, remove_duplicates)
 
-        self.ensemble.train(training_data, self.domain, self.featurizer, **kwargs)
+        self.ensemble.train(training_data, self.domain, self.featurizer,
+                            **kwargs)
 
-    def _prepare_training_data(self, filename, max_history, augmentation_factor,
+    def _prepare_training_data(self, resource_name, max_history,
+                               augmentation_factor,
                                max_training_samples=None,
                                max_number_of_trackers=2000,
                                remove_duplicates=True):
         """Reads training data from file and prepares it for the training."""
 
-        from rasa_core.training import extract_training_data_from_file
+        from rasa_core import training
 
-        if filename:
-            training_data = extract_training_data_from_file(
-                    filename,
+        if resource_name:
+            training_data = training.extract_training_data(
+                    resource_name,
                     self.domain,
                     self.featurizer,
                     interpreter=RegexInterpreter(),